Split API endpoints into different files
