--- data_store/loggers.py
@@ -1,27 +1,0 @@
-import sys
-import asyncio
-
-class ClientLogger:
-    def __init__(self):
-        self.message = ''
-        self.finished = False
-
-    async def log(self, value, end='\n'):
-        self.message += value + end
-        await asyncio.sleep(0)
-
-    def finish(self):
-        self.finished = True
-
-    async def iterate(self, startProcess):
-        await startProcess()
-        while not self.finished:
-            yield self.message
-            self.message = ''
-            await asyncio.sleep(0)
-        yield self.message
-        self.message = ''
-
-async def logToConsole(value, end='\n'):
-    sys.stderr.write('\x1b[0;32;40m' + value + end + '\x1b[0m')
-    sys.stderr.flush()
